Better string representation of relations
This might be a bit too complicated, but now there is a
GoString(Relation) function which shows the go string of the result of
a relation expression, and also a GoString() method of Relation, which
shows the Expression’s go string itself.  It might be better to be able
to convert relations to slice relations so that the string reps are all
unified and explicit?