Thanks Egon
Add Egon's suggestions to the todo list.  The amount of reflection scattered around the package is disconcerting.  If it were all unified by using a type for tuples instead of just interface{}, then it should be possible to limit it to just a few methods in a single file.  The intent of the code will also be much, much clearer.